backend/requirements.txt

fastapi==0.109.1
uvicorn==0.27.0
sqlalchemy==2.0.25
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.11.4
pydantic-settings==2.9.1
psycopg2-binary==2.9.10
boto3==1.34.34
botocore==1.34.34
